debug

+++ sys/printsyscallsummary.c
@@ -91,8 +91,8 @@ void printsyscallsummary() {
                     sum += durations[k];
                 }
                 kprintf("sum is %d, numcalls is %d\n", sum, sc[j].numcalls);
-                double avg = sum / sc[j].numcalls;
-                kprintf("\tSyscall: %s, count: %d, average execution time: %f (ms)\n", sc[j].name, sc[j].numcalls, avg);
+                double avg = (double)sum / sc[j].numcalls;
+                kprintf("\tSyscall: %s, count: %d, average execution time: %d (ms)\n", sc[j].name, sc[j].numcalls, avg);
             }
         }
     }
